inmates, tools: Add x86 linux loader
This simplifies the steps for booting Linux on x86 in a non-root cell.
It adds a jailhouse tool extension that builds the required parameters
for a minimalistic boot loader. So far the parameters consist of the
zero page and the kernel command line. Later on, we can also generate
ACPI tables.

The parameters can either be written to a file so that basic jailhouse
commands can be used on the target to create, load and start the cell.
Alternatively, the extension script can perform these steps itself,
enabling a single-line Linux cell boot.

Unfortunately, this comes with quite a bit manual translation of
driver/jailhouse.h into Python code. If the header should change, we
will have to adjust these bits.